refactor: fix guard imports
